Add test() to status_server. Move httpd to config.
